% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drpkg_skeleton-package.R
\name{drpkg_skeleton}
\alias{drpkg_skeleton}
\title{Package infrastructure for a mediation-testing toolkit}
\description{
This file provides package-level documentation and records import
requirements for functions in the \verb{R/} directory.
}
\keyword{internal}
