% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc-utils.R
\name{epsest.func}
\alias{epsest.func}
\title{Estimate the non-null proportion from Z-values}
\usage{
epsest.func(x, u, sigma)
}
\arguments{
\item{x}{A numeric vector of Z-values or other approximately
standardized test statistics.}

\item{u}{A numeric value specifying the null mean.}

\item{sigma}{A numeric value specifying the null standard deviation.}
}
\value{
A numeric scalar giving the estimated proportion of non-null signals.
}
\description{
Estimates the proportion of non-null effects in a large collection of
Z-statistics using an empirical characteristic-function method.
}
\details{
This function is an alternative implementation of the non-null
proportion estimator used in Jin--Cai style empirical-null methods. In
this package, it is mainly used as a helper for
\code{\link[=JC_prop_estmate]{JC_prop_estmate()}} when estimating mixture proportions from paired
Z-statistics.
}
\references{
Jin J, Cai TT (2007). \emph{Estimating the null and the proportion of
nonnull effects in large-scale multiple comparisons}. Journal of the
American Statistical Association, 102(478), 495--506.
}
