% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc-utils.R
\name{EstNull.func}
\alias{EstNull.func}
\title{Estimate empirical-null location and scale from Z-values}
\usage{
EstNull.func(x, gamma = 0.1)
}
\arguments{
\item{x}{A numeric vector of Z-values.}

\item{gamma}{A positive tuning parameter controlling the frequency
level used in empirical-null estimation. The default is
\code{0.1}.}
}
\value{
A list with components:
\describe{
\item{\code{mu}}{The estimated empirical-null mean.}
\item{\code{s}}{The estimated empirical-null standard deviation.}
}
}
\description{
Estimates the mean and standard deviation of an empirical null
distribution from a vector of Z-statistics using an empirical
characteristic-function approach.
}
\details{
This function provides an alternative implementation of empirical-null
estimation based on the empirical characteristic function. In this
package, it is mainly used by \code{\link[=JC_prop_estmate]{JC_prop_estmate()}} when estimating
component proportions from paired Z-statistics.

It is not part of the main TLR procedure proposed in the paper, but is
useful for comparison procedures and empirical-null based auxiliary
calculations.
}
\references{
Jin J, Cai TT (2007). \emph{Estimating the null and the proportion of
nonnull effects in large-scale multiple comparisons}. Journal of the
American Statistical Association, 102(478), 495--506.
}
