% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dact-utils.R
\name{nonnullPropEst}
\alias{nonnullPropEst}
\title{Estimate the proportion of non-null signals}
\usage{
nonnullPropEst(x, u, sigma)
}
\arguments{
\item{x}{A numeric vector of test statistics, typically Z-scores.}

\item{u}{A numeric value specifying the null mean.}

\item{sigma}{A numeric value specifying the null standard deviation.}
}
\value{
A numeric scalar giving the estimated proportion of non-null signals.
}
\description{
Estimates the proportion of non-null effects in a large collection of
test statistics using a Fourier-based method. This function can be used
to estimate marginal null proportions when constructing comparison
procedures for large-scale mediation testing.
}
\details{
The estimator is based on the empirical characteristic function and is
used here to estimate the marginal non-null proportion from a large set
of test statistics. In this package, it is primarily used as a
computational component in DACT-style combination rules and related
empirical-null adjustments.
}
\references{
Jin J, Cai TT (2007). \emph{Estimating the null and the proportion of
nonnull effects in large-scale multiple comparisons}. Journal of the
American Statistical Association, 102(478), 495--506.
}
