% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not-in}
\alias{not-in}
\alias{\%!in\%}
\title{Negation of \code{\%in\%}}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{A vector of values to match.}

\item{table}{A vector or list to match against.}
}
\value{
A logical vector.
}
\description{
Returns \code{TRUE} when \code{x} is not contained in \code{table}.
}
\examples{
1:3 \%!in\% c(2, 4)
}
