% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_temp_dir}
\alias{reset_temp_dir}
\title{Reset a temporary directory}
\usage{
reset_temp_dir(tempdir_path)
}
\arguments{
\item{tempdir_path}{Path to the temporary directory.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Deletes and recreates a temporary directory, then updates \code{TMPDIR} and the
\code{bigstatsr} temporary-directory option.
}
