% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{size_func}
\alias{size_func}
\title{Summarize size and power from a selected rejection set}
\usage{
size_func(sat_index, true_set, sim.num)
}
\arguments{
\item{sat_index}{An integer vector of indices selected by a testing
procedure.}

\item{true_set}{An integer vector of indices corresponding to the true
non-null hypotheses.}

\item{sim.num}{A positive integer giving the total number of simulated
hypotheses.}
}
\value{
If \code{true_set} is non-empty, returns a numeric vector of length two
containing:
\describe{
\item{size}{The empirical size, computed as the proportion of false
rejections among all hypotheses.}
\item{power}{The empirical power.}
}
If \code{true_set} is empty, returns the total number of false
positives.
}
\description{
Computes empirical size and power by comparing a selected set of
hypotheses with the true signal set in simulation studies.
}
\details{
This function is useful in simulation studies where one wants to report
size instead of false discovery rate. In the context of large-scale
mediation testing, it can be used to summarize the error-rate and power
performance of competing methods under the known data-generating
mechanism.
}
