% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlr.R
\name{TLR_pvalues}
\alias{TLR_pvalues}
\title{TLR p-values}
\usage{
TLR_pvalues(p.M, p.Y, p.M_one, p.Y_one, estws, coefs, coefs2)
}
\arguments{
\item{p.M, p.Y}{Marginal p-values used to estimate partial-null components.}

\item{p.M_one, p.Y_one}{P-values used in the TLR statistic.}

\item{estws}{Estimated null proportions.}

\item{coefs, coefs2}{Tail-approximation parameters.}
}
\value{
A numeric vector of p-values.
}
\description{
Computes p-values associated with the TLR statistic.
}
